beta app complete, need to finish settings features and notification.

+++ app/src/main/java/Entries/Entry.java
@@ -19,9 +19,9 @@ public class Entry {
 
     }
 
-    public Entry(String category, int amount){
+    public Entry(String category, float amount){
         this.category = category;
-        this.amount = amount;
+        this.amount = (double)amount;
     }
 
     public int getId() {
