+++ app/src/main/java/sqlite/SqlDAO.java
@@ -74,7 +74,7 @@ public class SqlDAO {
             int monthIndex = cursor.getColumnIndex("Month");
             int yearIndex = cursor.getColumnIndex("Year");
             DateC date = new DateC(cursor.getInt(dayIndex), cursor.getInt(monthIndex), cursor.getInt(yearIndex));
-            Entry entry = new Entry(cursor.getString(nameIndex), cursor.getInt(amountIndex), cursor.getString(categoryIndex), date);
+            Entry entry = new Entry(cursor.getString(nameIndex), cursor.getDouble(amountIndex), cursor.getString(categoryIndex), date);
             entry.setId(cursor.getInt(0));
 
             listOfEntries.add(entry);
